+++ RogueLike/Assets/Scripts/Enemy/EnemyRedHead.cs
@@ -48,12 +48,12 @@ public class EnemyRedHead : EntityStats
         _bonusLifetime += lifetime;
     }
 
-    public override void ChangeArmor(int armor)
+    public override void ChangeArmor(float armor)
     {
         
     }
 
-    public override void ChangeEvasion(int evasion)
+    public override void ChangeEvasion(float evasion)
     {
         
     }
